Rename file
The filename should match the only export
