+++ lib/prototypes/copy-prototype-methods.js
