+++ lib/prototypes/function.js
@@ -1,5 +1,5 @@
 "use strict";
 
-var copyPrototype = require("./copy-prototype");
+var copyPrototype = require("./copy-prototype-methods");
 
 module.exports = copyPrototype(Function.prototype);
